requirements.txt
streamlit
yfinance
pandas
numpy
matplotlib
google-generativeai
scikit-learn
tensorflow
openpyxl
xlsxwriter